requirements.txt
requests
python-decouple
fuzzywuzzy
python-Levenshtein
tabulate